pyproject.toml
[tool.poetry]
name = "nullify"
version = "0.1.0"
description = ""
authors = ["Ciaran <ciaran.finnegan@gmail.com>"]
readme = "README.md"
packages = [{include = "plugin"}]

[tool.poetry.dependencies]
python = "^3.9,<3.13"
cloudquery-plugin-sdk = "^0.1.6"
pydantic = "^2.4"
pyarrow = "^12.0"
requests = "^2.31.0"

[tool.poetry.scripts]
main = "main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

